+++ lib/generators/rspec/integration/integration_generator.rb
@@ -22,7 +22,7 @@ module Rspec
         WARNING
 
         template 'request_spec.rb',
-                 File.join('spec/requests', "#{name.underscore.pluralize}_spec.rb")
+                 target_path('requests', "#{name.underscore.pluralize}_spec.rb")
       end
     end
   end